requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
dnspython>=2.4.0
colorama>=0.4.6
tqdm>=4.66.0

